# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.LogInRequest:
    properties:
      email:
        maxLength: 40
        type: string
      password:
        maxLength: 40
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  auth.LogInWithTelegramRequest:
    properties:
      telegram_chat_id:
        minimum: 1
        type: integer
    required:
    - telegram_chat_id
    type: object
  auth.SignUpRequest:
    properties:
      email:
        maxLength: 40
        type: string
      password:
        maxLength: 40
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  auth.SignUpWithTelegramRequest:
    properties:
      full_name:
        maxLength: 40
        type: string
      telegram_chat_id:
        minimum: 1
        type: integer
    required:
    - full_name
    - telegram_chat_id
    type: object
  auth.TokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  auth.UserDetailsResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      notifications_enabled:
        type: boolean
      password_hash:
        type: string
      role:
        type: string
      telegram:
        type: integer
      user_id:
        type: integer
    type: object
  edu.BuildingResponse:
    properties:
      address:
        type: string
      building_id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
    type: object
  edu.FacultyResponse:
    properties:
      faculty_id:
        type: integer
      name:
        type: string
    type: object
  edu.ProgramResponse:
    properties:
      faculty_id:
        type: integer
      name:
        type: string
      program_id:
        type: integer
    type: object
  edu.TypeOfSubjectResponse:
    properties:
      name:
        type: string
      type_of_subject_id:
        type: integer
    type: object
...
  response.APIError:
    properties:
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: AntonioKrasava
  title: ClassFlow API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Аутентификация админ пользователя
      parameters:
      - description: Аутентификация пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.LogInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIError'
      summary: LogIn
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Создание нового админ пользователя
      parameters:
      - description: Создать пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIError'
      summary: SignUp
      tags:
      - auth
  /auth/telegram/login:
    post:
      consumes:
      - application/json
      description: Аутентификация студента
      parameters:
      - description: Аутентификация студента
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.LogInWithTelegramRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIError'
      summary: LogIn with telegram chat id
      tags:
      - auth
  /auth/telegram/signup:
    post:
      consumes:
      - application/json
      description: Создание студента с telegram chat id
      parameters:
      - description: Создать студента
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.SignUpWithTelegramRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIError'
      summary: SignUp with telegram chat id
      tags:
      - auth
  /auth/who:
    get:
      consumes:
      - application/json
      description: Получение информации о пользователе
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.UserDetailsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIError'
      security:
      - ApiKeyAuth: []
      summary: Who
      tags:
      - auth
  /edu/buildings:
    get:
      consumes:
      - application/json
      description: Получить список корпусов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/edu.BuildingResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIError'
      summary: GetAllBuildings
      tags:
      - edu
  /edu/faculties:
    get:
      consumes:
      - application/json
      description: Получить список всех факультетов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/edu.FacultyResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIError'
      summary: GetAllFaculties
      tags:
      - edu
  /edu/faculties/{faculty_id}/programs:
    get:
      consumes:
      - application/json
      description: Получить всех программ факультета
      parameters:
      - description: faculty ID
        in: path
        name: faculty_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/edu.ProgramResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIError'
      summary: GetProgramsByFacultyId
      tags:
      - edu
  /edu/types_of_subject:
    get:
      consumes:
      - application/json
      description: Получить список типов всех предметов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/edu.TypeOfSubjectResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIError'
      summary: GetAllTypesOfSubject
      tags:
      - edu
...
securityDefinitions:
  ApiKeyAuth:
    description: Use "Bearer <token>" to authenticate
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"